chore(zod): adds export for `makeLens`, adds test file
