+++ packages/zod/src/exports.ts
@@ -1,1 +1,2 @@
 export * from './version.js'
+export { makeLens } from './lens.js'
